+++ src/main/java/com/example/demo/serviceimplementation/MaterialConversionCalculateMaterialQuantityForServiceDoor.java
@@ -2,7 +2,7 @@ package com.example.demo.serviceimplementation;
 
 public class MaterialConversionCalculateMaterialQuantityForServiceDoor {
 
-    public static float materialConversionCalculateMaterialHisCode_11PLT003_00000(double DOOR_LENGTH, double DOOR_WIDTH) {
+    public static double materialConversionCalculateMaterialHisCode_11PLT003_00000(double DOOR_LENGTH, double DOOR_WIDTH) {
 
         return DOOR_LENGTH * 2 + DOOR_WIDTH;
